PAE-1590: Share one summary-log-submitted append primitive
The accredited path built the summary-log-submitted event inline while
registered-only emission had its own appendRegisteredOnlySubmittedEvent
helper — two places constructing the same event. Fold both onto a single
appendSummaryLogSubmittedEvent primitive that takes the accreditation
partition and credit total: the accredited path passes its computed total,
registered-only passes a null partition and zero. The reg-only repository
seam keeps its name; the duplication below it is gone.
